+++ CHANGELOG.md
@@ -31,6 +31,10 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
   If supporting these features causes problems, they may be removed again in a future version.
   [#123](https://github.com/hynek/stamina/pull/123)
 
+- An *on* hook can now return a float or a `datetime.timedelta` to specify a custom backoff that overrides the default backoff.
+  [#103](https://github.com/hynek/stamina/discussions/103)
+  [#125](https://github.com/hynek/stamina/pull/125)
+
 
 ### Fixed